detouch some of the utils from db/node to a separate file
